+++ apps/store/src/pages/_app.tsx
@@ -11,7 +11,7 @@ import globalCss from 'ui/src/global.css'
 import { theme } from 'ui'
 import { AppErrorDialog } from '@/components/AppErrorDialog'
 import { BankIdDialog } from '@/components/BankIdDialog'
-import { BankIdV6Dialog } from '@/components/BankIdV6Dialog'
+import { BankIdV6Dialog } from '@/components/BankIdV6Dialog/BankIdV6Dialog'
 import { ContactUs } from '@/components/ContactUs/ContactUs'
 import { GlobalLinkStyles } from '@/components/RichText/RichText.styles'
 import { usePublishWidgetInitEvent } from '@/features/widget/usePublishWidgetInitEvent'
